% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci_functions.R
\name{fibonacci_2}
\alias{fibonacci_2}
\title{Get the n-th Fibonacci's number}
\usage{
fibonacci_2(x)
}
\arguments{
\item{x}{non negative integer}
}
\value{
This function returns the n-th Fibonacci's number
}
\description{
This is a function that print the n-th Fibonacci's number
Change \code{x} to obtain the n-th Fibonacci's number
}
\examples{
fibonacci_2(1)

}
