% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris}
\alias{iris}
\title{Edgar Anderson's Iris Data}
\format{A data frame with columns:
\describe{
 \item{Sepal.Length}{Longitud del sépalo}
 \item{Sepal.Width}{Ancho del sépalo}
 \item{Petal.Length}{Longitud del pétalo}
 \item{Petal.Width}{Ancho del pétalo}
 \item{Species}{Especie de la flor}
}}
\source{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language.
 Wadsworth & Brooks/Cole.
}
\usage{
iris
}
\description{
Edgar Anderson's Iris Data
}
\examples{
\dontrun{
 iris
}
}
\keyword{datasets}
